dasi/requirements.txt
asgiref==3.7.2
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==4.3.2
click==8.1.7
cryptography==42.0.5
defusedxml==0.7.1
dj-rest-auth==2.2.8
Django==5.0.2
django-allauth==0.54.0
django-cors-headers==4.3.1
django-environ==0.11.2
django-rest-auth==0.9.5
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.7
environ==1.0
gunicorn==21.2.0
idna==3.6
inflection==0.5.1
joblib==1.3.2
JPype1==1.5.0
konlpy==0.6.0
lxml==5.2.1
mysqlclient==2.2.4
nltk==3.8.1
numpy==1.26.4
oauthlib==3.2.2
packaging==24.0
pillow==10.2.0
pycparser==2.21
PyJWT==2.8.0
PyMySQL==1.1.0
python3-openid==3.2.0
pytz==2024.1
PyYAML==6.0.1
regex==2023.12.25
requests==2.31.0
requests-oauthlib==1.3.1
rest-framework-simplejwt==0.0.2
scikit-learn==1.4.1.post1
scipy==1.13.0
six==1.16.0
sqlparse==0.4.4
threadpoolctl==3.4.0
tqdm==4.66.2
typing_extensions==4.10.0
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.1
whitenoise==6.6.0
